misc/HtmlTestSuite/results/test-0321.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>Compositing and Blending Level 1: parsing isolation with valid values</title>
<link href="mailto:ericwilligers@chromium.org" rel="author" title="Eric Willigers"/>
<link href="https://drafts.fxtf.org/compositing-1/#propdef-isolation" rel="help"/>
<meta content="isolation supports the full grammar '&lt;isolation-mode&gt;'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_valid_value("isolation", "auto");
test_valid_value("isolation", "isolate");
</script>
</body>
</html>
